refactor: group color fucntions
